requirements.txt
numpy==1.19.5
pytorch-lightning==1.2.5
tokenizers==0.10.3
#torch==1.12.0+cu113
tqdm==4.51.0
transformers==4.6.1
sentencepiece==0.1.97
sacrebleu==2.2.1     
tensorboardX==2.5.1 
dictdiffer==0.9.0  
tabulate==0.8.10 
termcolor==2.0.1 
torch-geometric==1.4.3
torch-sparse==0.6.15
torch-scatter==2.1.0